Freeze the date/time storage format and type the liveness cutoff
StdAdoDelegate's date/time and time-span conversions are non-virtual now. No
dialect ever overrode them, and the preferred-node liveness SQL does raw tick
arithmetic against LAST_CHECKIN_TIME and CHECKIN_INTERVAL, so an override
silently broke cluster failover for pinned triggers - which the store could
only detect by reflection and answer with a log line. UTC ticks and whole
milliseconds are stated as the schema contract instead, the reflective
override-detection machinery in JobStoreSupport is gone with the hazard it
watched for, and only the boolean pair remains a dialect seam. The
(DbDataReader, ordinal) reader-side reshape is deliberately not part of this
change.

TriggerAcquisitionCriteria.LiveNodeCutoff is a required DateTimeOffset like
its two time siblings. It was an optional long of UtcTicks, so omitting it
defaulted to zero - every node counted as dead, releasing pinned triggers to
anyone - and the unit was documented only in a SQL comment. The tick
conversion now lives in AddPreferredNodeParameters, which routes through the
frozen GetDbDateTimeValue so the binder and the SQL comment agree.
AddPagingParameters narrows skip/take from long to int, matching the
PagedQuery members it binds.

+++ src/Quartz/Impl/AdoJobStore/SQLiteDelegate.cs
@@ -40,7 +40,7 @@ public class SQLiteDelegate : StdAdoDelegate
     /// Binds the LIMIT/OFFSET parameters in the order the clause names them, which is the reverse of
     /// the ANSI clause's order and matters to providers that bind positionally.
     /// </summary>
-    protected override void AddPagingParameters(DbCommand cmd, long skip, long take, bool takeLimited)
+    protected override void AddPagingParameters(DbCommand cmd, int skip, int take, bool takeLimited)
     {
         if (takeLimited)
         {
